requirements.txt
python-telegram-bot=22.5
mcrcon=0.7.0